% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{data_formatting}
\alias{data_formatting}
\title{Format input dataset to be use by report_content()}
\usage{
data_formatting(file)
}
\arguments{
\item{file}{Path to the input file}
}
\value{
A csv
}
\description{
This function reformate input csv to be use by report_conten
}
