% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoBiofishr.R
\docType{package}
\name{faoBiofishr}
\alias{faoBiofishr}
\alias{faoBiofishr-package}
\title{Report compiler to biodiversity trent report of fisheries production}
\description{
Report compiler to biodiversity trent report of fisheries production
}
\details{
\tabular{ll}{
Package: \tab faoBiofishr\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2021-10-01\cr
LazyLoad: \tab yes\cr
}
}
\author{
Alexandre Bennici \email{bennicialexandre@gmail.com}
}
