% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_content.R
\name{report_content}
\alias{report_content}
\title{Compute content of report}
\usage{
report_content(file, out = NULL)
}
\arguments{
\item{file}{Path to the input file}

\item{out}{Path to export destination}
}
\value{
A list of element
}
\description{
This function compute all the output elements to include of report
}
