% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_execute.R
\name{report_execute}
\alias{report_execute}
\title{Execute report}
\usage{
report_execute(content, format = "pdf")
}
\arguments{
\item{content}{list of object produce by report_content()}

\item{format}{type of output}
}
\value{
report document
}
\description{
This function execute report
}
