% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_manager.R
\name{report_manager}
\alias{report_manager}
\title{Manage all porcess of producing report}
\usage{
report_manager(file, out = NULL, shiny.preview = FALSE, format = "pdf")
}
\arguments{
\item{file}{Path to the input file}

\item{out}{Path to export destination}

\item{shiny.preview}{display a preview of report with shiny (not yet implemented)}
}
\value{
report
}
\description{
This function is a report manager
}
