% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_shiny.R
\name{report_shiny}
\alias{report_shiny}
\title{Produce Shiny preview}
\usage{
report_shiny(content)
}
\arguments{
\item{content}{list of object produce by report_content()}
}
\value{
shiny application
}
\description{
This function provide shiny preview of report to reedit it (not yef finished)
}
